tests: 64 bit compatibility
Fixed compilation errors. Mostly DEBUG/printf formatting and void pointer casting.

Other changes are:
* net/gnrc_sixlowpan_frag_*: Generalized packet size calculation
* cpu/native_backtrace: Reduced required backtrace size to 3 for 64-bit
* periph/flashpage: Simplified test
* unittests/tests-pktbuf: Generalized alignment
* sys/architecture: Extended test for 64-bit

+++ tests/drivers/at86rf2xx/main.c
@@ -73,7 +73,7 @@ int netdev_ieee802154_minimal_init_devs(netdev_event_cb_t cb) {
     puts("Initializing AT86RF2XX devices");
 
     for (unsigned i = 0; i < AT86RF2XX_NUM; i++) {
-        printf("%d out of %d\n", i + 1, AT86RF2XX_NUM);
+        printf("%d out of %u\n", i + 1, (unsigned)AT86RF2XX_NUM);
         /* setup the specific driver */
         at86rf2xx_setup(&at86rf2xx[i], &at86rf2xx_params[i], i);
